\name{biogeophyl-package}
\alias{biogeophyl-package}
\alias{biogeophyl}
\docType{package}
\title{
\packageTitle{biogeophyl}
}
\description{
\packageDescription{biogeophyl}
}
\details{
\packageDESCRIPTION{biogeophyl}
\packageIndices{biogeophyl}
}
\author{
\packageAuthor{biogeophyl}

Maintainer: \packageMaintainer{biogeophyl}
}
\references{
Martins, W. S.; Carmo, W. C.; Longo, H. J.; Rosa, T. C. & Rangel, T. F. (2013) SUNPLIN: Simulation with Uncertainty for Phylogenetic Investigations. \emph{BMC Bioinformatics}, \bold{14}, 324
}
\keyword{ package }
\seealso{
\code{\link[ape:ape-package]{ape}}, \code{\link[phytools:phytools-package]{phytools}}
}
