\name{biogeophyl-package}
\alias{biogeophyl-package}
\alias{biogeophyl}
\docType{package}
\title{
\packageTitle{biogeophyl}
}
\description{
\packageDescription{biogeophyl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{biogeophyl}
\packageIndices{biogeophyl}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{biogeophyl}

Maintainer: \packageMaintainer{biogeophyl}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
